added groupmasters and related functionalities

+++ Server/ConcordiaCurriculumManager/Controllers/GroupController.cs
@@ -4,6 +4,7 @@ using ConcordiaCurriculumManager.Services;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Authorization;
 using Swashbuckle.AspNetCore.Annotations;
+using ConcordiaCurriculumManager.Security;
 
 namespace ConcordiaCurriculumManager.Controllers;
 
@@ -13,10 +14,14 @@ namespace ConcordiaCurriculumManager.Controllers;
 public class GroupController : Controller
 {
     private readonly IGroupService _groupService;
+    private readonly IUserAuthenticationService _userService;
+    private readonly ILogger<CourseController> _logger;
 
-    public GroupController(IGroupService groupService)
+    public GroupController(IGroupService groupService, IUserAuthenticationService userService, ILogger<CourseController> logger)
     {
         _groupService = groupService;
+        _userService = userService;
+        _logger = logger;
     }
 
     [HttpPost(nameof(CreateGroup))]
@@ -65,11 +70,34 @@ public class GroupController : Controller
     }
 
     [HttpPost("{groupId}/users/{userId}")]
-    [Authorize(Roles = RoleNames.Admin)]
+    [Authorize]
     [SwaggerResponse(StatusCodes.Status200OK, "User added to group successfully")]
     [SwaggerResponse(StatusCodes.Status404NotFound, "Error adding user to group")]
     public async Task<IActionResult> AddUserToGroup(Guid groupId, Guid userId)
     {
+        Guid currentUserId;
+        try
+        {
+            currentUserId = Guid.Parse(_userService.GetCurrentUserClaim(Claims.Id));
+        }
+        catch (Exception e)
+        {
+            var message = "Unexpected error occured while retrieving user information.";
+            _logger.LogWarning($"${message}: {e.Message}");
+            return Problem(
+                title: message,
+                detail: e.Message,
+                statusCode: StatusCodes.Status500InternalServerError);
+        }
+
+        bool isAdmin = User.IsInRole(RoleNames.Admin);
+        bool isMaster = await _groupService.IsGroupMaster(currentUserId, groupId);
+
+        if (!isAdmin && !isMaster)
+        {
+            return StatusCode(StatusCodes.Status403Forbidden, "You are not authorized to add users to this group");
+        }
+
         var result = await _groupService.AddUserToGroup(userId, groupId);
         if (result)
         {
@@ -79,12 +107,72 @@ public class GroupController : Controller
     }
 
     [HttpDelete("{groupId}/users/{userId}")]
-    [Authorize(Roles = RoleNames.Admin)]
+    [Authorize]
     [SwaggerResponse(StatusCodes.Status200OK, "User removed from group successfully")]
     [SwaggerResponse(StatusCodes.Status404NotFound, "Error removing user from group")]
     public async Task<IActionResult> RemoveUserFromGroup(Guid groupId, Guid userId)
     {
+        Guid currentUserId;
+        try
+        {
+            currentUserId = Guid.Parse(_userService.GetCurrentUserClaim(Claims.Id));
+        }
+        catch (Exception e)
+        {
+            var message = "Unexpected error occured while retrieving user information.";
+            _logger.LogWarning($"${message}: {e.Message}");
+            return Problem(
+                title: message,
+                detail: e.Message,
+                statusCode: StatusCodes.Status500InternalServerError);
+        }
+
+        bool isAdmin = User.IsInRole(RoleNames.Admin);
+        bool isMaster = await _groupService.IsGroupMaster(currentUserId, groupId);
+
+        if (!isAdmin && !isMaster)
+        {
+            return StatusCode(StatusCodes.Status403Forbidden, "You are not authorized to add users to this group");
+        }
+
         var result = await _groupService.RemoveUserFromGroup(userId, groupId);
+        if (result)
+        {
+            return Ok();
+        }
+        return NotFound();
+    }
+
+    [HttpPost("{groupId}/masters/{userId}")]
+    [Authorize]
+    [SwaggerResponse(StatusCodes.Status200OK, "User added to group masters successfully")]
+    [SwaggerResponse(StatusCodes.Status404NotFound, "User is not a member of this group or is already a group master")]
+    public async Task<IActionResult> AddGroupMaster(Guid groupId, Guid userId)
+    {
+        Guid currentUserId;
+        try
+        {
+            currentUserId = Guid.Parse(_userService.GetCurrentUserClaim(Claims.Id));
+        }
+        catch (Exception e)
+        {
+            var message = "Unexpected error occured while retrieving user information.";
+            _logger.LogWarning($"${message}: {e.Message}");
+            return Problem(
+                title: message,
+                detail: e.Message,
+                statusCode: StatusCodes.Status500InternalServerError);
+        }
+
+        bool isAdmin = User.IsInRole(RoleNames.Admin);
+        bool isMaster = await _groupService.IsGroupMaster(currentUserId, groupId);
+
+        if (!isAdmin && !isMaster)
+        {
+            return StatusCode(StatusCodes.Status403Forbidden, "You are not authorized to add users to this group");
+        }
+
+        var result = await _groupService.AddGroupMaster(userId, groupId);
         if (result)
         {
             return Ok();
